Isolated LineChart From Home Fragment
